# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRVM API Documentation",
        "version": "1.0.0",
        "description": "API documentation for MyReverseVendingMachine Server",
        "contact": {
            "email": "admin@myrvm.com"
        }
    },
    "servers": [
        {
            "url": "/",
            "description": "Default Server (Relative)"
        }
    ],
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user and return token",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "MyDevice"
                                    }
                                }
                            },
                            "examples": {
                                "Super Admin": {
                                    "value": {
                                        "email": "superadmin@myrvm.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "Admin": {
                                    "value": {
                                        "email": "admin@myrvm.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "Operator": {
                                    "value": {
                                        "email": "operator@myrvm.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "Technician": {
                                    "value": {
                                        "email": "tech@myrvm.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "Tenant": {
                                    "value": {
                                        "email": "tenant@starbucks.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "User 1": {
                                    "value": {
                                        "email": "john@example.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "User 2": {
                                    "value": {
                                        "email": "jane@example.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                },
                                "User 3": {
                                    "value": {
                                        "email": "bob@example.com",
                                        "password": "password123",
                                        "device_name": "MyDevice"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid credentials"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "logoutUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful logout"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/vouchers": {
            "get": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Get list of active vouchers (Public)",
                "operationId": "getVouchersList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/logs": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Read system logs (Admin Only)",
                "operationId": "getSystemLogs",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "YYYY-MM-DD",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log content"
                    },
                    "302": {
                        "description": "Redirect - Unauthorized access redirected to login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin Role Required"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/rvm-ui/{machine_uuid}": {
            "get": {
                "tags": [
                    "Kiosk UI"
                ],
                "summary": "Kiosk Interface (Web View)",
                "description": "Main touchscreen interface entry point.",
                "parameters": [
                    {
                        "name": "machine_uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Machine Serial Number (e.g. RVM-GI-001)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML Page"
                    },
                    "404": {
                        "description": "Machine Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}